{"schema":"decodex/commit/1","summary":"Modularize macOS capture host contracts","authority":"manual"}
